Add files via upload
Changed layout.html and StyleSheet to look different. Added a class called DBConnection so you only have to type in the connection string once, to achieve this changed a small block of code on all the .cshtml.cs pages just before conn.open(), to make this work for the database just change the conenction string on the DBConnection class.